+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+##2017/08/18
+
+- add code to return home page on click the logo icon.
+
 ##2017/08/13
 
 - show date on canvas clock